requirements.txt
requests~=2.31.0
mysql-connector-python~=8.2.0
Flask~=3.0.0
python-dotenv~=1.0.0
geopy~=2.4.1
Flask-Cors~=4.0.0